+++ spec/overcommit/hook/commit_msg/message_format_spec.rb
@@ -1,6 +1,6 @@
 require 'spec_helper'
 
-describe Overcommit::Hook::CommitMsg::MsgPattern do
+describe Overcommit::Hook::CommitMsg::MessageFormat do
   let(:config)  { Overcommit::ConfigurationLoader.default_configuration }
   let(:context) { double('context') }
   subject { described_class.new(config, context) }
@@ -20,23 +20,27 @@ describe Overcommit::Hook::CommitMsg::MsgPattern do
     let(:config) do
       super().merge(Overcommit::Configuration.new(
         'CommitMsg' => {
-          'MsgPattern' => {
+          'MessageFormat' => {
             'pattern' => nil
           }
         }
       ))
     end
+
     let(:commit_msg) { 'Some Message' }
+
     it { should pass }
   end
 
   context 'when message does not match the pattern' do
     let(:commit_msg) { 'Some Message' }
+
     expected_message = [
       'Commit message pattern mismatch.',
       'Expected : <Issue Id> | <Commit Message Description> | <Developer(s)>',
       'Sample : DEFECT-1234 | Refactored Onboarding flow | John Doe'
     ].join("\n")
+
     it { should fail_hook expected_message }
   end
 end