+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 * Improve error message output when there is a problem processing messages
   via `extract_messages` pre-commit hook helper
 * Fix `Minitest` pre-push hook to include all test files
+* Add `MessageFormat` commit-msg hook to validate commit messages against
+  a regex pattern
 
 ## 0.32.0.rc1